move language status service to workbench
